+++ test/native/src/test/java/org/apache/shardingsphere/test/natived/commons/algorithm/ClassBasedInlineShardingAlgorithmFixture.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.natived.jdbc.commons.algorithm;
+package org.apache.shardingsphere.test.natived.commons.algorithm;
 
 import org.apache.shardingsphere.sharding.api.sharding.standard.PreciseShardingValue;
 import org.apache.shardingsphere.sharding.api.sharding.standard.RangeShardingValue;